segmentation/requirements.txt
acvl-utils==0.2
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
albucore==0.0.19
albumentations==1.4.20
alembic==1.13.2
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
batchgenerators==0.25
batchgeneratorsv2==0.2.1
bottle==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
comm==0.2.2
conda-pack==0.8.0
connected-components-3d==3.18.0
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dicom2nifti==2.4.11
dipy==1.7.0
docker-pycreds==0.4.0
dynamic-network-architectures==0.3.1
efficientnet-pytorch==0.7.1
einops==0.8.0
entrypoints==0.4
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.0.1
fft-conv-pytorch==1.2.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
greenlet==3.0.3
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
imagecodecs==2024.6.1
imageio==2.35.1
importlib_metadata==8.2.0
importlib_resources==6.4.3
intensity-normalization==2.2.4
ipykernel==6.29.5
ipyparallel==8.8.0
ipython==8.18.1
itk==5.4.0
itk-core==5.4.0
itk-filtering==5.4.0
itk-io==5.4.0
itk-numerics==5.4.0
itk-registration==5.4.0
itk-segmentation==5.4.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.6
linecache2==1.0.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
monai==1.3.2
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==3.2.2
nnunetv2==2.5.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
optuna==3.6.1
optuna-dashboard==0.16.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pretrainedmodels==0.7.4
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
pydicom==2.4.4
Pygments==2.18.0
pymedio==0.2.14
pyparsing==3.1.2
pystackreg==0.2.7
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.1
requests==2.32.3
safetensors==0.4.4
scikit-fuzzy==0.4.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
segmentation-models-pytorch==0.3.3
sentry-sdk==2.13.0
setproctitle==1.3.3
shapely==2.0.5
SimpleITK==2.4.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.32
stack-data==0.6.3
statsmodels==0.14.2
stringzilla==3.10.5
sympy==1.13.2
threadpoolctl==3.5.0
tifffile==2024.8.10
timm==0.9.2
tomli==2.0.1
torch==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traceback2==1.4.0
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
unittest2==1.1.0
urllib3==2.2.2
wandb==0.17.7
wcwidth==0.2.13
yacs==0.1.8
yarl==1.9.4
zipp==3.20.0
